Fill out equipment inventory section.
