+++ src/megameklab/com/printing/PrintMech.java
@@ -421,144 +421,6 @@ public class PrintMech extends PrintEntity {
                 || (loc == Mech.LOC_CLEG);
     }
     
-    @Override
-    protected void writeEquipment(SVGRectElement svgRect) {
-        Map<Integer, Map<RecordSheetEquipmentLine,Integer>> eqMap = new TreeMap<>();
-        Map<String,Integer> ammo = new TreeMap<>();
-        for (Mounted m : mech.getEquipment()) {
-            if ((m.getType() instanceof AmmoType)
-                    && (((AmmoType) m.getType()).getAmmoType() != AmmoType.T_COOLANT_POD)) {
-                if (m.getLocation() != Entity.LOC_NONE) {
-                    String shortName = m.getType().getShortName().replace("Ammo", "");
-                    shortName = shortName.replace("(Clan)", "");
-                    String munition = ((AmmoType) m.getType()).getSubMunitionName().replace("(Clan) ", "");
-                    shortName = shortName.replace(munition, "");
-                    ammo.merge(shortName.trim(), m.getBaseShotsLeft(), Integer::sum);
-                }
-                continue;
-            }
-            if ((m.getType() instanceof AmmoType)
-                    || (m.getLocation() == Entity.LOC_NONE)
-                    || !UnitUtil.isPrintableEquipment(m.getType(), true)) {
-                continue;
-            }
-            if (mech.hasETypeFlag(Entity.ETYPE_QUADVEE)
-                    && (m.getType() instanceof MiscType)
-                    && m.getType().hasFlag(MiscType.F_TRACKS)) {
-                continue;
-            }
-            eqMap.putIfAbsent(m.getLocation(), new HashMap<>());
-            RecordSheetEquipmentLine line = new RecordSheetEquipmentLine(m);
-            eqMap.get(m.getLocation()).merge(line, 1, Integer::sum);
-        }
-        
-        Rectangle2D bbox = getRectBBox(svgRect);
-        Element canvas = (Element) svgRect.getParentNode();
-        int viewWidth = (int)bbox.getWidth();
-        int viewHeight = (int)bbox.getHeight();
-        int viewX = (int)bbox.getX();
-        int viewY = (int)bbox.getY();
-        
-        int qtyX = (int) Math.round(viewX + viewWidth * 0.037);
-        int nameX = (int) Math.round(viewX + viewWidth * 0.08);
-        int locX = (int) Math.round(viewX + viewWidth * 0.41);
-        int heatX = (int) Math.round(viewX + viewWidth * 0.48);
-        int dmgX = (int) Math.round(viewX + viewWidth * 0.53);
-        int minX = (int) Math.round(viewX + viewWidth * 0.75);
-        int shortX = (int) Math.round(viewX + viewWidth * 0.82);
-        int medX = (int) Math.round(viewX + viewWidth * 0.89);
-        int longX = (int) Math.round(viewX + viewWidth * 0.96);
-        
-        int indent = (int) Math.round(viewWidth * 0.02);
-        
-        int currY = viewY + 10;
-        
-        float fontSize = FONT_SIZE_MEDIUM;
-        float lineHeight = getFontHeight(fontSize) * 1.2f;
-        
-        addTextElement(canvas, qtyX, currY, "Qty", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, nameX + indent, currY, "Type", fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, locX,  currY, "Loc", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, heatX, currY, "Ht", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, dmgX, currY, "Dmg", fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, minX, currY, "Min", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, shortX, currY, "Sht", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, medX, currY, "Med", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        addTextElement(canvas, longX, currY, "Lng", fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_BOLD_VALUE);
-        currY += lineHeight * 1.2;
-
-        int lines = 0;
-        for (Integer loc : eqMap.keySet()) {
-            for (RecordSheetEquipmentLine line : eqMap.get(loc).keySet()) {
-                int rows = line.nRows();
-                if ((rows == 1) && (getTextLength(line.getNameField(0, mech.isMixedTech()), fontSize) > locX - nameX)) {
-                    rows++; 
-                }
-                lines += rows;
-            }
-        }
-        if (lines > 11) {
-            lineHeight = getFontHeight(fontSize) * 1.0f;
-        }
-        if (lines > 13) {
-            lineHeight = getFontHeight(fontSize) * 0.8f;
-        }
-        if (lines > 16) {
-            fontSize = FONT_SIZE_SMALL;
-        }
-        if (lines > 20) {
-            fontSize = FONT_SIZE_VSMALL;
-        }
-        
-        for (Integer loc : eqMap.keySet()) {
-            for (RecordSheetEquipmentLine line : eqMap.get(loc).keySet()) {
-                for (int row = 0; row < line.nRows(); row++) {
-                    if (row == 0) {
-                        addTextElement(canvas, qtyX, currY, Integer.toString(eqMap.get(loc).get(line)), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                        lines = addMultilineTextElement(canvas, nameX, currY, locX - nameX - indent, lineHeight,
-                                line.getNameField(row, mech.isMixedTech()), fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-
-                    } else {
-                        lines = addMultilineTextElement(canvas, nameX + indent, currY, locX - nameX - indent, lineHeight,
-                                line.getNameField(row, mech.isMixedTech()), fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    }
-                    addTextElement(canvas, locX,  currY, line.getLocationField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    addTextElement(canvas, heatX, currY, line.getHeatField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    lines = Math.max(lines, addMultilineTextElement(canvas, dmgX, currY, minX - dmgX - fontSize, lineHeight,
-                            line.getDamageField(row), fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_NORMAL_VALUE));
-                    addTextElement(canvas, minX, currY, line.getMinField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    addTextElement(canvas, shortX, currY, line.getShortField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    addTextElement(canvas, medX, currY, line.getMediumField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    addTextElement(canvas, longX, currY, line.getLongField(row), fontSize, SVGConstants.SVG_MIDDLE_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-                    currY += lineHeight * lines;
-                }
-            }
-        }
-        
-        String quirksText = formatQuirks();
-
-        if ((ammo.size() > 0) || (quirksText.length() > 0)) {
-            Element svgGroup = getSVGDocument().createElementNS(svgNS, SVGConstants.SVG_G_TAG);
-            canvas.appendChild(svgGroup);
-            lines = 0; 
-            if (ammo.size() > 0) {
-                lines = addMultilineTextElement(svgGroup, viewX + viewWidth * 0.025, 0, viewWidth * 0.95, lineHeight,
-                        "Ammo: " + ammo.entrySet().stream()
-                        .map(e -> String.format("(%s) %d", e.getKey(), e.getValue()))
-                        .collect(Collectors.joining(", ")), fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-            }
-            if (quirksText.length() > 0) {
-                lines += addMultilineTextElement(svgGroup, viewX + viewWidth * 0.025, lines * lineHeight,
-                        viewWidth * 0.95, lineHeight,
-                        "Quirks: " + quirksText, fontSize, SVGConstants.SVG_START_VALUE, SVGConstants.SVG_NORMAL_VALUE);
-            }
-            svgGroup.setAttributeNS(null, SVGConstants.SVG_TRANSFORM_ATTRIBUTE,
-                    String.format("%s(0,%f)", SVGConstants.SVG_TRANSLATE_VALUE,
-                            viewY + viewHeight - lines * lineHeight));
-        }        
-
-    }
-    
     private void writeLocationCriticals(int loc, SVGRectElement svgRect) {
         Rectangle2D bbox = getRectBBox(svgRect);
         Element canvas = (Element) svgRect.getParentNode();
